On WSL modify setup script

+++ setup_dependency.zsh
@@ -1,6 +1,10 @@
 #!/bin/zsh
 #oh-my-zsh
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+if [[ ! -e "$HOME/.oh-my-zsh" ]]; then
+    sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+else
+    echo "Oh-My-Zsh has already installed."
+fi
 #zsh-autosuggestions
 git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
 #zsh-syntax-highlighting
@@ -10,5 +14,9 @@ git clone --depth=1 https://gitee.com/romkatv/powerlevel10k.git ${ZSH_CUSTOM:-$H
 #vundle.vim
 git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
 #vifm-colors
-rm -rf ~/.config/vifm/colors
-git clone https://github.com/vifm/vifm-colors ~/.config/vifm/colors
+if [[ ! -e "$HOME/.config/vifm/colors/ph.vifm" ]]; then
+    rm -rf ~/.config/vifm/colors
+    git clone https://github.com/vifm/vifm-colors ~/.config/vifm/colors
+else
+    echo "Vifm colorshemes already exist."
+fi
